fix scalars mismatch

+++ crates/coordinator/src/api.rs
@@ -19,7 +19,7 @@ pub struct Health {
 }
 
 pub async fn health(State(state): State<AppState>) -> Json<Health> {
-    let db_ok = sqlx::query_scalar::<_, i64>("SELECT 1")
+    let db_ok = sqlx::query_scalar::<_, i32>("SELECT 1")
         .fetch_one(&state.pool)
         .await
         .is_ok();
